Working, starting migration of adapters into LogStream where they belong
